+++ include/tape/ITape.hpp
@@ -20,8 +20,8 @@ public:
   virtual bool AtLeftBound() const = 0;
   virtual bool AtRightBound() const = 0;
 
-  virtual size_t Length() const = 0;
-  virtual size_t Position() const = 0;
+  virtual uint64_t Length() const = 0;
+  virtual uint64_t Position() const = 0;
 
 public:
   // Following are helper functions which must use only operations defined above
